requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9
drf-yasg>=1.21
djangorestframework-simplejwt>=5.3
